+++ aws/ecs-application/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
       fallback on default value.
     * `enabled_datadog_sidecar`, `dd_site`, `dd_api_key_arn`, `dd_agent_image`, `dd_port`: supports datadog sidecar
       definitions.
+    * `use_alb`: whether to use ALB.
 
 ## [0.1.78]() (2025-03-10)