+++ aws/ecs-application/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.2]() (2025-03-18)
+
+### Features
+
+* Add ECS Service Connect using `enabled_service_connect` and `service_connect_configuration`.
+
 ## [0.2.1]() (2025-03-18)
 
 ### Features